Import live.2004.06.14.tar.gz
- Updated "MPEG1or2Demux" to make the checking for a PACK_START_CODE more
  robust if it's not immediately present where we expect to see it.
- Fixed a bug in "JPEGVideoRTPSource" that could cause a problem on
  64-bit architectures.
  (Thanks to Andrey Filippov for noticing this.)

+++ liveMedia/JPEGVideoRTPSource.cpp
@@ -409,7 +409,7 @@ Boolean JPEGVideoRTPSource
 
     unsigned hdrlen = computeJPEGHeaderSize(qtlen, dri);
     resultSpecialHeaderSize -= hdrlen; // goes negative
-    headerStart += resultSpecialHeaderSize; // goes backward
+    headerStart += (int)resultSpecialHeaderSize; // goes backward
     createJPEGHeader(headerStart, type, width, height, qtables, qtlen, dri);
   }
